requirements.txt
astroid==2.2.5
azure-cognitiveservices-speech==1.3.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
entrypoints==0.3
filelock==3.0.10
flake8==3.7.7
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
flake8-single-quotes==0.1.0
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-speech==1.0.0
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
llvmlite==0.28.0
logzero==1.5.0
mccabe==0.6.1
numba==0.43.1
numpy==1.16.2
pbr==5.2.0
pluggy==0.9.0
protobuf==3.7.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
resampy==0.2.1
rsa==4.0
scipy==1.2.1
six==1.12.0
snowballstemmer==1.2.1
toml==0.10.0
tornado==6.0.1
tox==3.9.0
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.5.0
watson-developer-cloud==2.10.1
websocket-client==0.48.0
wrapt==1.11.1
